+++ app.py
@@ -1,231 +1,146 @@
-import requests
-import json
-import os
+from flask import Flask, request, jsonify, render_template, redirect,url_for,flash
+from flask_sqlalchemy import SQLAlchemy
+from werkzeug.security import generate_password_hash,check_password_hash
+from flask_login import LoginManager, UserMixin, login_user, login_required, logout_user,current_user
 import webbrowser
-from threading import Timer,Thread,Event,Lock
-from dotenv import load_dotenv
-from utils.prompt import prompts
-from flask import Flask, render_template, request, jsonify,send_from_directory
-import_lock=Lock()
-def import_functions():
-    global functions
-    with import_lock:
-         print("importing functions and neccessary libraries.........\n")
-         import functions
-         print("the functions successfully imported\n")
-import_thread=Thread(target=import_functions)
-import_thread.start()
-print("starting the voice assistant......\n")
-import_thread.join()
+from utils.config import Config
+from utils.rate_limiter import RateLimiter
+from utils.gemini_client import GeminiClient
+from utils.file_service import FileService
+
+
 app = Flask(__name__, static_folder='static', template_folder='templates')
+app.config['SECRET_KEY'] = 'group2'  # Change this to a secure secret key
+app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///users.db'  # Using SQLite for simplicity
+db = SQLAlchemy(app)
+login_manager = LoginManager()
+login_manager.init_app(app)
+login_manager.login_view = 'login'
+# Initialize services
+class User(UserMixin, db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    email = db.Column(db.String(120), unique=True, nullable=False)
+    contact = db.Column(db.String(10), unique=True, nullable=False)
+    password_hash = db.Column(db.String(128))
+    name = db.Column(db.String(100), nullable=False)
 
-class Backend:
-    def __init__(self):
-        load_dotenv()
-        self.api_key = os.getenv("GOOGLE_API_KEY")
-        self.system_message_flag = ''
-        self.stop_event=Event()
-        self.messages=[]
-        self.messages1=[]
-        self.backup_messages=[]
-        self.headers = {
-    'Content-Type': 'application/json'
-      }
-    def count_tokens(self,message):
-        try:
-            data1={
-                    "contents":[message]
-                }
-            count_tokens_response=requests.post("https://generativelanguage.googleapis.com/v1beta/models/gemini-pro:countTokens?key="+self.api_key,json=data1,headers=self.headers)
-            count_token=count_tokens_response.json()
-            print(count_token)
-            total_tokens=count_token.get('totalTokens')
-            print(total_tokens)
-            return total_tokens
-        except Exception as e:
-             print(f"unable to count tokens reason :{e}")
-    def write_list_to_file(self, lst, filename):
-        with open(filename, 'a', encoding='utf-8') as file:
-            for item in lst:
-                file.write(str(item) + '\n')
+    def set_password(self, password):
+        self.password_hash = generate_password_hash(password)
 
-    def parse_function_response(self,function_name,function_args):
-        print("Gemini: Called function " + function_name )
+    def check_password(self, password):
+        return check_password_hash(self.password_hash, password)
 
-        try:
-            arguments = function_args
-            print("arguments:",arguments)
-            if hasattr(functions, function_name):
-                function_response = getattr(functions, function_name)(**arguments)
-            else:
-                function_response = "ERROR: Called unknown function"
-        except TypeError as e:
-            function_response = f"ERROR: Invalid arguments:{e}"
-        fmessage = {
-                "role": "function",
-                "parts": [
-                    {
-                        "functionResponse": {
-                            "name": function_name,
-                            "response": {
-                                "name": function_name,
-                                "content": function_response
-                            }
-                        }
-                    }
-                ]
-            }
-            
-            # Save the function response to the JSON file and append to messages
-        with open("messages.json", "a") as f:
-                f.write(json.dumps(fmessage, indent=4))
-        self.messages.append(fmessage)
-    
-
-        return self.process_message()
-
-    
-    def gemini(self):
-        count_tokens=0
-        data = {
-            "contents": [self.messages],
-            "tools": [{
-                "functionDeclarations": functions.function_definitions
-            }
-
-            ],
-            "safetySettings": [
-                {
-                    "category": "HARM_CATEGORY_DANGEROUS_CONTENT",
-                    "threshold": "BLOCK_ONLY_HIGH"
-                },
-                {
-                    "category": "HARM_CATEGORY_HARASSMENT",
-                    "threshold": "BLOCK_ONLY_HIGH"
-
-                },
-                {
-                    "category": "HARM_CATEGORY_HATE_SPEECH",
-                    "threshold": "BLOCK_ONLY_HIGH",
-                },
-                {
-                    "category": "HARM_CATEGORY_SEXUALLY_EXPLICIT",
-                    "threshold": "BLOCK_LOW_AND_ABOVE",
-                },
-            ],
-           "generationConfig": {
-            "temperature": 0.3
-        }
-            
-        }
-        # count_tokens_response=requests.post("https://generativelanguage.googleapis.com/v1beta/models/gemini-pro:countTokens?key="+self.api_key,json=data1,headers=headers)
-        
-        if(self.count_tokens(self.messages)<30720):
-            response = requests.post("https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-flash:generateContent?key=" +self.api_key, json=data)
-            if response.status_code != 200:
-                print(f"errorcode:200->{response.text}")
-                if response.status_code==500:
-                    self.messages=[]
-                    return f"an iternal server has occured on google please try reloading the page."
-                if response.status_code==400:
-                    return f"an  internal server error please make sure the requsest is between user and model"
-                return f"unable to send a request to the gemini server please reload the page to try again"
-        else:
-               messag=self.summarize_content(self.messages)
-               return f"reduced content after the summarization:{messag}"
-        response = response.json()
-        if "candidates" not in response:
-            return "an internal error has occured .something with the prompt handling.contact developer to solve this error."
-        if "content" not in response["candidates"][0]:
-            message={
-                "role":"model",
-                "parts":"i am sorry i am unable to generate the response"
-            }
-            self.messages.append(message)
-            return f"unable to  generated  the response.try again. "
-
-        message = response['candidates'][0]['content']['parts']
-        message1={
-                "role": "model",
-                "parts": message
-            }
-        with open("messages.json", "a") as f:
-                f.write(json.dumps(message1, indent=4))
-        self.messages.append(message1)
-        return message
-    def process_message(self):
-        text_content = None
-        function_name = None
-        function_args = None
-        message=self.gemini()
-        # Parse through each part in the message to handle text and function call
-        for part in message:
-            if "text" in part:
-                text_content = part["text"]
-            if "functionCall" in part:
-                function_name = part["functionCall"].get("name")
-                function_args = part["functionCall"].get("args")
-    
-        # Process function call if present
-        if function_name:
-            # Execute function call
-            return self.parse_function_response(function_name, function_args)
-        return text_content  # Return only text if present
+@login_manager.user_loader
+def load_user(user_id):
+    return User.query.get(int(user_id))
 
 
-    def call_gemini(self, system_message=None):
-        if system_message:
-            message = {"role": "user", "parts": [{"text": system_message}]}
-            with open("messages.json", "a") as f:
-                f.write(json.dumps(message, indent=4))
-            self.messages=[]
-            self.messages.append(message)
-        response = self.process_message()
-        if response:
-            print("bot:", response)
-            self.system_message_flag = response
-        else:
-            self.system_message_flag = ''
+config = Config()
+rate_limiter = RateLimiter(config.MAX_RPM, config.MAX_TPM, config.MAX_RPD)
+gemini_client = GeminiClient(config.GOOGLE_API_KEY, rate_limiter)
+file_service = FileService(config.UPLOAD_FOLDER, config.ALLOWED_EXTENSIONS)
 
-    def send_system_message(self):
-        if self.system_message_flag:
-            message = self.system_message_flag
-            self.system_message_flag = ''
-            print("Sending it to the client")
-            return jsonify({'message': message,'messages':self.messages})
-        else:
-            return jsonify({'message': '','messages':self.messages})
 
-    def send_message(self):
-        print("Send message endpoint is called")
-        data = request.json
+@login_manager.user_loader
+def load_user(user_id):
+    return User.query.get(int(user_id))
+
+@app.route('/')
+def home():
+    return render_template('index.html')
+@app.route('/send_message', methods=['POST'])
+def send_message():
+    try:
+        data = request.get_json()
+        print("data:",data)
         message = data.get('message', '')
+        
         if not message:
-            print("No message is provided by the server")
-        message = {
-                "role": "user",
-                "parts": [{"text": message}]
-            }
-        with open("messages.json", "a") as f:
-                f.write(json.dumps(message, indent=4))
-        self.messages.append(message)
-        bot_response = self.process_message()
-        print(f"sending to the server:{bot_response}")
-        return jsonify({'message': bot_response,'messages':self.messages})
-    def send_pdf(self,message):
-        with open("messages.json", "a") as f:
-                f.write(json.dumps(message, indent=4))
-        self.messages.append(message)
-        response=self.process_message()
-        print(f"sending to the server:{response}")
-        return jsonify({'message':response,'messages':self.messages})
-    def restart_server(self):
-        print("restarting")
-        self.messages=self.backup_messages
-        bot_response = self.gemini()
-        # self.speak_async(bot_response)
-        return jsonify({'message': bot_response,'messages':self.messages})
+            return jsonify({"error": "Message is required"}), 400
+            
+        # Get recent conversation history
+        
+        # Send message to Gemini
+        response = gemini_client.send_message(message)
+        
+        # Save to memory
+        
+        return jsonify({
+            "message": response
+        })
+        
+    except Exception as e:
+        print("error:",e)
+        return jsonify({"error": str(e)}), 500
+@app.route('/register', methods=['GET', 'POST'])
+def register():
+    if request.method == 'POST':
+        name = request.form['name']
+        email = request.form['email']
+        contact = request.form['contact']
+        password = request.form['password']
+        confirm_password = request.form['confirm_password']
 
+        # Validation
+        if User.query.filter_by(email=email).first():
+            flash('Email already registered!')
+            return redirect(url_for('register'))
+        
+        if User.query.filter_by(contact=contact).first():
+            flash('Contact number already registered!')
+            return redirect(url_for('register'))
+
+        if password != confirm_password:
+            flash('Passwords do not match!')
+            return redirect(url_for('register'))
+
+        # Create new user
+        user = User(email=email, contact=contact, name=name)
+        user.set_password(password)
+        
+        try:
+            db.session.add(user)
+            db.session.commit()
+            flash('Registration successful! Please login.')
+            return redirect(url_for('login'))
+        except Exception as e:
+            db.session.rollback()
+            flash('An error occurred during registration.')
+            return redirect(url_for('register'))
+
+    return render_template('register.html')
+@app.route('/login', methods=['GET', 'POST'])
+def login():
+    if request.method == 'POST':
+        email = request.form.get('email')
+        password = request.form.get('password')
+        
+        user = User.query.filter_by(email=email).first()
+        
+        if user is None:
+            # New user registration
+            flash("user not found !! please register")
+            return redirect(url_for("register"))
+        else:
+            # Existing user login
+            if not user.check_password(password):
+                flash('Invalid password!')
+                return redirect(url_for('login'))
+            
+        login_user(user)
+        return redirect(url_for('input'))
+    
+    return render_template('login.html')
+
+@app.route("/input")
+@login_required
+def input():
+    return render_template("input.html",user_name=current_user.name)
+@app.route('/logout')
+@login_required
+def logout():
+    logout_user()
+    return redirect(url_for('home'))
 def check_env_file():
     try:
         with open('.env', 'r') as f:
@@ -241,6 +156,7 @@ def check_env_file():
 def check_keys():
     keys_present = check_env_file()
     return jsonify({'keys_present': keys_present})
+
 @app.route('/save_keys', methods=['POST'])
 def save_keys():
     data=request.get_json()
@@ -251,57 +167,9 @@ def save_keys():
         f.write(f"""TF_ENABLE_ONEDNN_OPTS=0\n""")
 
     return jsonify({'success': True})
-
-
-
-backend = Backend()
-
-@app.route('/')
-def home():
-    return render_template('index.html')
-@app.route('/no_camera_input')
-def no_camera_input():
-    system_message =prompts[1]['start_without_camera']
-    backend.call_gemini(system_message=system_message)
-    return render_template("input.html")
-@app.route('/input')
-def input():
-    system_message =prompts[0]['start_with_camera']
-    backend.call_gemini(system_message=system_message)
-    return render_template('input.html')
-@app.route('/upload', methods=['POST'])
-def upload_file():
-    if 'file' not in request.files:
-        return  jsonify({"message":"no file part"})
-    file = request.files['file']
-    if file.filename == '':
-        return 'No selected file'
-    if file and file.filename.endswith('.pdf'):
-        # Save the file to a desired location
-        if not os.path.exists("documents"):
-             os.makedirs("documents")
-        file.save(os.path.join('documents', file.filename))
-        print(file.filename)
-        message = {
-                "role": "user",
-                "parts": [{"text": f"summarize the {file.filename} file"}]
-            }
-        return  backend.send_pdf(message) 
-    else:
-        return 'File upload failed: Only PDF files are allowed'
 @app.route('/documentation')
 def documentation():
     return render_template('documentation.html')
-
-@app.route('/resources')
-def resources():
-    return render_template('references.html')
-@app.route('/images')
-def show_images():
-    output_dir='output'
-    if not os.listdir(output_dir):
-        return jsonify({'error': 'Output directory is empty'})
-    return send_from_directory()
 @app.route('/about')
 def about():
     return render_template('about.html')
@@ -310,20 +178,33 @@ def about():
 def contact():
     return render_template('contact.html')
 
-@app.route('/send_system_message', methods=['POST'])
-def send_system_message():
-    return backend.send_system_message()
+@app.route('/resources')
+def resources():
+    return render_template('references.html')
+@app.route('/upload', methods=['POST'])
+def upload_file():
+    try:
+        if 'file' not in request.files:
+            return jsonify({"error": "No file part"}), 400
+            
+        file = request.files['file']
+        filepath = file_service.save_file(file)
+        
+        message = f"summarize the {file.filename} file"
+        response = gemini_client.send_message(message)
+        
+        return jsonify({"message": response})
+        
+    except ValueError as e:
+        return jsonify({"error": str(e)}), 400
+    except Exception as e:
+        return jsonify({"error": str(e)}), 500
 
-@app.route('/send_message', methods=['POST'])
-def send_message():
-    return backend.send_message()
-
-@app.route('/restart_backend', methods=['POST'])
-def restart_backend():
-    return backend.restart_server()
 def open_browser():
     webbrowser.open('http://localhost:5000')
 
 if __name__ == "__main__":
-    Timer(1, open_browser).start()
-    app.run(debug=True,host= "0.0.0.0",port="5000")
+    # Timer(1, open_browser).start()
+    with app.app_context():
+        db.create_all()
+    app.run(host="0.0.0.0", debug=True, port="5000")
